[Add] DeletePlaylistsTest - a_user_cannot_delete_albums_from_other_uses

+++ app/Http/Controllers/AlbumController.php
@@ -38,6 +38,8 @@ class AlbumController extends Controller
 
     public function destroy(Album $album)
     {
+        $this->authorize('delete', $album);
+
         $album->delete();
 
         return response()->json([], 204);
